perf: prerender /brands/[slug] and /flavor/[id] at build time
Add generateStaticParams + dynamicParams = false so all 92 brand pages
and ~5.3k flavor pages are statically generated. The flavor route was
previously a client-side fetch; split it into a server page (direct
shishaData read) + a client wrapper for framer-motion. Unknown slugs/ids
now 404 at build-manifest level instead of running through the
notFound() branch on every request.
